fix: fix unit tests for get_more opreation
